pyproject.toml
[project]
name = "autoeditor"
version = "0.1.0"
description = "Automatic video editing based on script matching"
requires-python = ">=3.12"
dependencies = [
    "beartype>=0.22.7",
    "elevenlabs>=1.50.0",
    "nicegui>=3.4.1",
    "numpy>=2.0.0",
    "openai>=1.68.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "pywebview>=6.1",
    "scipy>=1.14.0",
]

[project.scripts]
autoeditor = "src.main:main"
autoeditor-ui = "src.ui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pyright>=1.1.407",
]
